requirements-flitsers-dev.txt
Flask==1.0.2
Flask-SQLAlchemy==2.3.2
Jinja2==2.10.1
psycopg2==2.7.5
SQLAlchemy==1.3.3
python-dotenv==0.10.1
